Refactor(layout): 내브바 있는 레이아웃 리팩토링
- 레이아웃 컴포넌트명이 헷갈리게 되어 있어서 수정하고 폴더 분리
- 핸들러 함수명 오타 수정
