+++ src/components/features/layout/nav-bar/index.tsx
@@ -1,24 +1,48 @@
-import NavBarButtons from './buttons.tsx';
-import { Box } from '@chakra-ui/react';
+import { useNavigate } from 'react-router-dom';
 
-type NavBarTypes = {
-  children: React.ReactNode;
-};
+import { navBarData } from '@/constants/nav-bar';
+import { Box, Image } from '@chakra-ui/react';
+import styled from '@emotion/styled';
 
-const Index = ({ children }: NavBarTypes) => {
+const NavBar = () => {
+  const navigate = useNavigate();
+
+  const handleNav = (page: string) => {
+    navigate(page);
+  };
+
   return (
-    <>
-      <Box
-        display='flex'
-        flexDirection='column'
-        height='100%'
-        paddingBottom='3.44'
-      >
-        <Box flex={1}>{children}</Box>
-        <NavBarButtons />
-      </Box>
-    </>
+    <Wrapper>
+      {navBarData.map((item) => (
+        <Box width='2rem' height='2rem' flexShrink={0} key={item.title}>
+          <Image
+            onClick={() => handleNav(item.path)}
+            src={item.icon}
+            alt={item.title}
+          />
+        </Box>
+      ))}
+    </Wrapper>
   );
 };
 
-export default Index;
+export default NavBar;
+
+const Wrapper = styled(Box)`
+  display: flex;
+  flex-direction: row;
+
+  position: fixed;
+  width: 100%;
+  bottom: 0;
+  left: 0;
+
+  height: 3.44rem;
+
+  justify-content: center;
+  align-items: center;
+  gap: 2.5rem;
+
+  border-top: 0.5px solid #bdc5cd;
+  background-color: #fafafa;
+`;
